# src/utils/swagger-output-user.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "TELL API Documentation",
    "description": "Technology Enabled Language Learning"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1/user",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "x-auth-token",
      "description": "Authentication token"
    }
  },
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "create account for new user",
        "description": "Add Desc Here",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Register",
            "in": "body",
            "description": "Req.Body Data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "fullName": {
                  "type": "string",
                  "example": "Sethu K"
                },
                "email": {
                  "type": "string",
                  "example": "sethu@gmail.com"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "9876543210"
                },
                "password": {
                  "type": "string",
                  "example": "zA1@sethu"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registered Successfully."
          },
          "400": {
            "description": "Validaiton Error."
          },
          "409": {
            "description": "Account Already Exists."
          },
          "500": {
            "description": "Server Error."
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "user login with their credentials",
        "description": "Add Desc Here",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Login",
            "in": "body",
            "description": "Req.Body Data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "sethu@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "zxcASD123@!$"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login Successful."
          },
          "400": {
            "description": "Validaiton Error."
          },
          "401": {
            "description": "Incorrect Email or Password."
          },
          "404": {
            "description": "Account Not Exists."
          },
          "500": {
            "description": "Server Error."
          }
        }
      }
    },
    "/markResult": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "overall mark & speed for the level",
        "description": "Add Desc Here",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Result",
            "in": "body",
            "description": "Req.Body Data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "QA_ID": {
                  "type": "string",
                  "example": "6555c642fe134b47d29f820f"
                },
                "questionResult": {
                  "type": "string",
                  "example": "how are you feeling right now"
                },
                "answerResult": {
                  "type": "string",
                  "example": "I am feeling fine thank you"
                },
                "questionMark": {
                  "type": "number",
                  "example": 80
                },
                "answerMark": {
                  "type": "number",
                  "example": 70
                },
                "level": {
                  "type": "number",
                  "example": 1
                },
                "timeTakenForQuestion": {
                  "type": "number",
                  "example": 5
                },
                "timeTakenForAnswer": {
                  "type": "number",
                  "example": 3
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result Submitted Successfully or Level Completed."
          },
          "400": {
            "description": "Validaiton Error."
          },
          "401": {
            "description": "Invalid token."
          },
          "404": {
            "description": "Question Not Found."
          },
          "500": {
            "description": "Server Error."
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "User & Admin"
        ],
        "summary": "Logs out current logged in user session",
        "description": "Add Desc Here",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Logged out Successful."
          },
          "401": {
            "description": "Invalid token."
          },
          "500": {
            "description": "Server Error."
          }
        }
      }
    }
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}